Extract daily mode logic into separate module
